Plateforme-de-Gestion-des-Examens-Universitaires-BDA-main/requirements.txt
streamlit==1.30.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
mysql-connector-python==8.3.0
SQLAlchemy==2.0.23
